+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 - No longer filters OWL:Ontology Subject when building 'subjectsOnly()'. Need
   to re-examine why this was ever needed (see comment in code).
+- Added SKOS:prefLabel to the collection of terms applicable for labels.
 
 ### Deprecation notice